Whac a mole Game
whac-a-mole

Vanilla JavaScript game

Learn to make a simple grid-based game using HTML, CSS and JavaScript. The idea of whack-a-mole is simple! The player needs to hit the grid with the mole in as many times as possible until the time runs out.

    document.querySelector
    textContent
    forEach
    addEventListener
    setInterval
    classList.add

...and much more.

View the source code

As always, if you get stuck or have any questions, please do comment below and I will be sure to get back to you!

I would love to see what you made so please tag me on twitter on @Mudassir Ali or mention me on Linkedin so I can find it and see.